% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createEvent}
\alias{createEvent}
\title{Function to call when creating an event}
\usage{
createEvent(start, end, name = "Untitled", location = "", participants_id)
}
\arguments{
\item{start}{start date and hour of the event (Uses the POSIXct format)}

\item{end}{end date and hour of the event (Uses the POSIXct format)}

\item{name}{name of the event (String)}

\item{location}{name of the location (For now a string in the future a GPS)}
}
\description{
Function to call when creating an event
}
