% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{list_to_event}
\alias{list_to_event}
\title{Create an event based on a list input of form : start,end,name,location,participants_id(string)}
\usage{
list_to_event(list)
}
\description{
Create an event based on a list input of form : start,end,name,location,participants_id(string)
}
\details{
Please use only in dev function
}
