% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\docType{class}
\name{Person-class}
\alias{Person-class}
\title{Class describing a person in the package}
\description{
Class describing a person in the package
}
\section{Slots}{

\describe{
\item{\code{person_id}}{An identifier used to differenciate persons}
}}

