% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{printGantt}
\alias{printGantt}
\title{Function used to display a gantt calendar of the tasks}
\usage{
printGantt(calendar)
}
\arguments{
\item{calendar}{the dataframe to be displayed}
}
\description{
Function used to display a gantt calendar of the tasks
}
\details{
Still cut the last task at hand
}
